+++ administration_logic/trunk/src/java/main/com/orpheus/administration/handlers/SponsorImageApprovalRejectionHandler.java
@@ -3,10 +3,6 @@
  */
 package com.orpheus.administration.handlers;
 
-import javax.servlet.http.HttpServletRequest;
-
-import org.w3c.dom.Element;
-
 import com.orpheus.administration.Helper;
 import com.orpheus.administration.persistence.AdminData;
 import com.orpheus.game.persistence.Domain;
@@ -16,6 +12,9 @@ import com.topcoder.user.profile.UserProfile;
 import com.topcoder.web.frontcontroller.ActionContext;
 import com.topcoder.web.frontcontroller.Handler;
 import com.topcoder.web.frontcontroller.HandlerExecutionException;
+import org.w3c.dom.Element;
+
+import javax.servlet.http.HttpServletRequest;
 
 /**
  * Provides an abstract base handler implementation that approves or rejects a
@@ -275,7 +274,7 @@ abstract class SponsorImageApprovalRejectionHandler implements Handler {
         if (sponsorId == null) {
             return failedResult;
         }
-        if (domain.getSponsorId() != sponsorId.longValue()) {
+        if (domain.getSponsorId().longValue() != sponsorId.longValue()) {
             Helper.processFailureSponsorNotBelongToDomain(request,
                     failRequestAttrName, sponsorId.longValue(), domainId
                             .longValue());
